refactor: remove monumentsCount update logic from administrative division and monument services
