+++ territory/src/main/java/pt/estga/territory/repositories/AdministrativeDivisionRepository.java
@@ -2,13 +2,11 @@ package pt.estga.territory.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import pt.estga.territory.entities.AdministrativeDivision;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
@@ -33,23 +31,4 @@ public interface AdministrativeDivisionRepository extends JpaRepository<Administ
 
     @Query("SELECT d FROM AdministrativeDivision d WHERE d.osmAdminLevel = :adminLevel AND d.monumentsCount > 0")
     List<AdministrativeDivision> findWithMonuments(@Param("adminLevel") int adminLevel);
-
-    @Modifying
-    @Query("UPDATE AdministrativeDivision d SET d.monumentsCount = d.monumentsCount + 1 WHERE d.id = :divisionId")
-    void incrementMonumentsCount(@Param("divisionId") Long divisionId);
-
-    @Modifying
-    @Query("UPDATE AdministrativeDivision d SET d.monumentsCount = d.monumentsCount - 1 WHERE d.id = :divisionId AND d.monumentsCount > 0")
-    void decrementMonumentsCount(@Param("divisionId") Long divisionId);
-
-    @Modifying
-    @Query(value = """
-        UPDATE administrative_division d
-        SET monuments_count = (
-            SELECT COUNT(m.id)
-            FROM monument m
-            WHERE m.parish_id = d.id OR m.municipality_id = d.id OR m.district_id = d.id
-        )
-    """, nativeQuery = true)
-    void recalculateAllMonumentsCounts();
 }
